requirements.txt
# Core Data Processing
numpy==1.24.3
pandas==2.0.3

# Machine Learning / Deep Learning
tensorflow==2.15.0
scikit-learn==1.3.2

# Streaming & Big Data
kafka-python==2.0.2
pyspark==3.5.0

# Backend / API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2

# Dashboard
streamlit==1.29.0
altair==5.2.0
watchdog==3.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
joblib==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Development Tools
jupyter==1.0.0
ipykernel==6.27.1
black==23.11.0
flake8==6.1.0

# Optional: Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Optional: Cloud Storage
boto3==1.33.2
